Fix a bad istio maste merge issue, causing the test to fail. (#2131)
Automatic merge from submit-queue.

Fix a bad master merge issue, causing a test to fail.

Before submitting the PR that adds this test, I did a bad merge with changes in Master. Coincidentally the test was also not included in the bazel BUILD file and therefore presubmit didn't catch it.
thanks @mandarjog for finding this issue.

**What this PR does / why we need it**:

**Which issue this PR fixes** *(optional, in `fixes #<issue number>(, fixes #<issue_number>, ...)` format, will close that issue when PR gets merged)*: fixes #

**Special notes for your reviewer**:

**Release note**:

```release-note
none
```