Fix data races #1945 (#2439)
Automatic merge from submit-queue.

Fix data races #1945

This PR attempts to fix the data races highlighted in #1945 

All of the data races fixed here were data races in the tests themselves except for one in the mock. All have been resolved by serialising access to the counters with a mutex.

There is one more data race listed in #1945 (TestSyncer) but that seems to be in the k8s client code rather than in the istio code. The test also looks like more of an integration type test. Any guidance would be appreciated?