Get golint working & speed up each invocation of make by a second (#2575)
Automatic merge from submit-queue.

Get golint working & speed up each invocation of make by a second

In the continuous build the "running golint on packages..." is followed by a long sequence of:

/bin/bash: line 1: golint: command not found

So, this change does something akin to dep and has make download golint when needed and specify a full path when executing golint.

The next issue is that golint complains about not finding libraries.  Essentially passing "./..." to "go list" causes it to output absolute paths that are prefixed with an underscore.  Setting GOPATH when calling "go list" resolves this.  I'm aware that the istio docs prescribe that people set this in advance but the core build already makes an attempt to cope without this in order to make the build more turnkey and self-sufficient.

With this change golint complains about 3 dozen items, though they all seemed cosmetic in nature (vs causing an issue in runtime behavior) so I punted on fixing them.

I also added in a change from my go version check change that avoids shelling out to run various apps (including a "find") on each invocation of make that seems to take about 1.2 seconds on my system.