Log on stray go scheduled routines (#2816)
Automatic merge from submit-queue.

Log on stray go scheduled routines

Log if go routine are not properly closed by adapters after they have been closed() on the handler interface.

Based on offline discussion with @ozevren I have also made a change where now the CloseFailure counter is reported on the old snapshot config instead of new snapshot config. This is done because miss behaving handler is actually tied to the old config and not the one which the mixer has already flipped to.