Revert optimization, breaks re-running. Convert to unix. (#2907)
Automatic merge from submit-queue.

Revert optimization, breaks re-running. Convert to unix.

Sorry about the big diff - but I can't make my editor to use windows line ending, and 
all the files are and should be using \n.

This is urgent, prevents re-running ( when re-run happens, there is no workspace ). In general
workspace can act as a cache, but we should not depend on it. We'll probably need to iterate
few more times - but golang 1.10 may make remove the need for this optimization.