Fix docker tar generation for release builds (#3376)
Automatic merge from submit-queue.

Fix docker tar generation for release builds

This change allows release builds to complete successfully.

The baseline docker targets used to build images as $(IMAGE) (which implies $(IMAGE):latest) and yesterday a change was committed that changed this to $(HUB)/$(IMAGE):$(TAG) but didn't update the tar-generation rule to reflect this name change.  This change primarily updates the tar-generation rule to reflect the change.

The release build doesn't care about the HUB (it just wants the tar files), but I can't set an empty hub (the Makefile doesn't permit one, but even if it did the image name would probably end up with a '/' prefix).  The Makefile's default hub is "istio", and this name is also now baked into the tar so I had to update release/store_artifacts.sh and release/publish_release.sh to reflect this change.

I also changed release/cloud_builder.sh to specify a HUB.  This is just to help insulate the build from any change to the default HUB in the top-level Makefile.  Although "docker.io/istio" is specified the docker tool seems to implicitly change this to just "istio".

I tested that a release build can complete successfully with this change, though I did no testing of the artifacts to see if there's any 2nd order side effects.