Disable vm test in cluster wide (#3613)
Automatic merge from submit-queue.

Disable vm test in cluster wide

Cluster provisioning is enabled on cluster wide test, and vm test has not been updated to use the pre-provisioned VM and is creating instead a VM in the main testing project. The VM test is using local network IP, and the cluster and the VM are not in the same project which makes communication between those 2 impossible.

Disabling the vm test for now until we can take advantage of the provisioned VM