should not create k8s client if config dir is provided (#3779)
Automatic merge from submit-queue.

should not create k8s client if config dir is provided

in response to istio-users question here: https://groups.google.com/forum/#!topic/istio-users/52SiQkkTq2A

pilot should not need to create a k8s client (and therefor provide a kubeconfig file) when the `--configDir` flag is provided.  This is currently blocking the use of pilot's filewatcher adapter.  The way we are currently checking for whether a k8s client needs to be created has a bug and needs to be corrected - this will be addressed in an additional PR