Add a logic to remove stale clusters (#6410)
When an istio service with an associated endpoint is removed, Pilot seem to
have kept the stale cluster(s) in StreamAggregatedResources, part of the problem
was the initial check was only looking for the length equality of the clusters
from service discovery and the ones that Envoy knows about. This change performs
a deep equality check on both cluster sets. If both sets are equal no
changes made, otherwise we refresh clusters that envoy knows about with
the new set.