Add support for a check cache in Mixer. (#5753)
This cache is intended to serve as a large level 2 cache to supplement the proxy's much smaller level 1
cache. The point is to try to eliminate as many calls to Check adapters, and hence check backends, as
possible.

The max # entries in the cache can be adjusted from the command-line. Default is a fairly beefy
1,500,000 entries, which is sufficient for supporting a 5000 QPS rate with a TTL of 5 minute
per entry. We can tune the default after experience to see what works best.

This doesn't do anything for quotas. Quota requests must still flow to the backends unconditionally.