Unify reconnection and message processing in mcp client. (#7811)
* Unify retry and processing logic in mcp client.

Merge the (re)connection and processing logic into a single function
and unify the error handling and retry code paths. Update the ctrlz
tests to account for the fact that the initial MCP requests are now
captured in the journel.

This also adds unit tests for the reconnection logic on send and
receive failures.

* remove extra debug code

* fix race in pkg/mcp/configz unit tests

* linter fixes

* update comments

* add deleted line