CircleCI gate for Test Framework w/ Kubernetes environment.  (#8735)
* Add Istio deployment support to the Test Framework.

This change adds a very basic deployment support for deploying Istio
on Kubernetes to the Test Framework. It is based on Helm Template
generation, and the core infrastructure allows using a variety of Helm
value files, as well as allowing the specification of values piece-meal.

The wire-up to the actual Test Framework is pretty anemic right now,
and that is intentional. As we make more progress on the dependency/setup
layer of the Test Framework, those should drive how the integration works.

* Accommodate Lab => CI naming change.

* Skip Galley Validator tests that are also skipped in E2E.

* Cleanup and refactoring.

* More refactoring.

* Make WorkDir be based on RunID.

* Minor fix.

* Add more documentation.

* Make linter happy.

* Disable collection of delete errors during close.

* Fix leftover parameter after rename.

* Accommodate CR feedback.

* Oh mighty lint, who gates our checkins! Forgive us and let me pass!

* Wait for the main namespace deletion in deployment.Delete.

* Lint gods are angry again! They demand another CR as a sacrifice!