Istio cni e2e (#9577)
* Add a CNI specific e2e test

This change adds a e2e simple test with
the CNI plugin used instead of initContainers.

* e2e_simple prow check

This commit changes the test to be based
off the e2e_simple that prow runs.

* Address review comments

* Adds a hook for prow to execute a CNI test

This changes adds a shell script that triggers
the e2e_simple test to run with the CNI enabled.

* Simplified the e2e_simple_cni target

* Use the gke resource with networkpolicy enabled

* Remove some debugging